requirements.txt
accelerate
bitsandbytes
torch
transformers
einops
sentencepiece
chainlit
python-dotenv
